+++ SecureAccessBLE/Classes/Tracking/TrackingEvent.swift
@@ -11,6 +11,11 @@ import Foundation
 internal enum TrackingEvent: String {
     case interfaceInitialized
 
+    case bluetoothPoweredON
+    case bluetoothPoweredOFF
+    case bluetoothUnauthorized
+    case bluetoothUnsupported
+
     case discoveryStartedByApp
     case discoveryStarted
     case discoveryCancelledbyApp
@@ -55,6 +60,11 @@ internal enum TrackingEvent: String {
              .serviceGrantRequestFailed,
              .serviceGrantResponseReceived:
             return "ServiceGrant"
+        case .bluetoothPoweredON,
+             .bluetoothPoweredOFF,
+             .bluetoothUnsupported,
+             .bluetoothUnauthorized:
+            return "Bluetooth"
         }
     }
 
@@ -94,11 +104,23 @@ internal enum TrackingEvent: String {
             return "Failure in requesting service grant"
         case .serviceGrantResponseReceived:
             return "Service grant response is received"
+        case .bluetoothPoweredON:
+            return "Bluetooth is powered on"
+        case .bluetoothPoweredOFF:
+            return "Bluetooth is powered off"
+        case .bluetoothUnauthorized:
+            return "Bluetooth is unauthorized"
+        case .bluetoothUnsupported:
+            return "Bluetooth is unsupported"
         }
     }
 
     var shouldBeReportedToTacs: Bool {
-        return [TrackingEvent.connectionTransferringBLOB].contains(self)
+        return [TrackingEvent.connectionTransferringBLOB,
+                TrackingEvent.bluetoothPoweredOFF,
+                TrackingEvent.bluetoothPoweredON,
+                TrackingEvent.bluetoothUnauthorized,
+                TrackingEvent.bluetoothUnsupported].contains(self)
     }
 
     var defaultParameters: [String: Any] {